+++ packages/javascript/bh-shared-ui/src/views/PrivilegeZones/Details/MembersList.tsx
@@ -21,7 +21,8 @@ import { NodeIcon, SortableHeader } from '../../../components';
 import { InfiniteQueryFixedList, InfiniteQueryFixedListProps } from '../../../components/InfiniteQueryFixedList';
 import { SortOrder } from '../../../types';
 import { cn } from '../../../utils';
-import { SelectedHighlight, getListHeight } from './utils';
+import { getListHeight } from './utils';
+import { SelectedHighlight } from './SelectedHighlight';
 
 interface MembersListProps {
     listQuery: UseInfiniteQueryResult<{
